Range check on item sound effects.  Fixed float to int warning.
git-svn-id: svn://localhost/media/c/Documnts/src/flare/repo.svndump/trunk@61 6ae5efe3-03d3-4ed7-8a64-3d71375a9a67

+++ src/MapCollision.cpp
@@ -73,7 +73,7 @@ bool MapCollision::line_check(int x1, int y1, int x2, int y2, int checktype) {
 	float dy = abs(y2 - y1);
 	float step_x;
 	float step_y;
-	int steps = max(dx, dy);
+	int steps = (int)max(dx, dy);
 	int mapx;
 	int mapy;
